+++ src/Stripe.net/Services/Invoices/ShippingOptions.cs
@@ -1,0 +1,26 @@
+// File generated from our OpenAPI spec
+namespace Stripe
+{
+    using Newtonsoft.Json;
+
+    public class ShippingOptions : INestedOptions
+    {
+        /// <summary>
+        /// Customer shipping address.
+        /// </summary>
+        [JsonProperty("address")]
+        public AddressOptions Address { get; set; }
+
+        /// <summary>
+        /// Customer name.
+        /// </summary>
+        [JsonProperty("name")]
+        public string Name { get; set; }
+
+        /// <summary>
+        /// Customer phone (including extension).
+        /// </summary>
+        [JsonProperty("phone")]
+        public string Phone { get; set; }
+    }
+}
